requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
numpy>=1.24.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
requests>=2.31.0
tqdm>=4.66.0